Add files via upload
A pokedex... or so they call it.
I actually thought it was an actual pokedex, where you open somehow choose a pokemon and it will maybe open the data of that pokemon..
I will do that in an actual challenge, though. As this will set up most things for that. It uses an API linked to pokemon - pretty cool.